Permission add and delete comments

+++ comments/views.py
@@ -1,20 +1,24 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 from rest_framework.permissions import IsAuthenticated
 
 from .serializers.common import CommentSerializer
 from .models import Comment
 
 
+  # ***** ADD A COMMENT ******
+
 # /comments/
 class CommentListView(APIView):
     permission_classes = (IsAuthenticated, )
   
-  # ***** CREATE A COMMENT ******
-
     def post(self, request):
+        print("REQUEST USER ->", request.user)
+        request.data['owner'] = request.user.id
+        print("REQUEST DATA ->", request.data)
+      
         try:
             comment_to_add = CommentSerializer(data=request.data)
             if comment_to_add.is_valid():
@@ -29,9 +33,15 @@ class CommentListView(APIView):
 class CommentDetailView(APIView):
     permission_classes = (IsAuthenticated, )
 
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         try:
             comment_to_delete = Comment.objects.get(pk=pk)
+            
+            print('FOUND COMMENT OWNER', comment_to_delete.owner)
+            print('REQUEST USER', request.user)
+            if comment_to_delete.owner != request.user:
+                raise PermissionDenied('Unauthorised')
+              
             comment_to_delete.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except Comment.DoesNotExist as e:
